Link to teensy4-rs RTIC examples from README
mciantyre/teensy4-rs#97 migrated all of the project's RTIC examples to
RTIC 1.0. This commit adds a link to the project for those seeking more
examples. Closes #8.

teensy4-rs and its associated projects are in flux, so it's easier for
me to maintain the examples when they're separate from the RTIC examples
repo.