Add Example for an IWDG on an STM32F4 with RTIC v1.0
This adds a simple example for how an IWDG can be used with RTIC on an
STM32F4 microcontroller. It has been tested against an STM32F401RE CPU
on a Nucleo F401RE development board with the built-in user button
connected on PC13.

This has been tested with stable Rust 1.65.0, it is unknown which older
versions of Rust also support this, but the support is determined by the
used libraries.

The `CHANGELOG.md` has been slightly extended to align it more closely
with the "keep a changelog" standard which it tries to adhere to. This
might need updating in case some of the entries have already been
released (are there even releases of this repository? If not, the file
doesn't make too much sense and the commit history should IMHO suffice).

Solves rtic-rs/rtic-examples#42