add dependabot dependency version update config
this enables the [dependabot dependency version update][].
this means that dependabot will weekly check if there are updates
available and will then raise PRs with these updates (if they can be
applied in an automated way). this helps in ensuring that the
dependencies are kept up to date.

it's important that new projects are also added to the config, otherwise
they won't be covered by this.
i'm not aware of any way to tell dependabot that it should just look for
`Cargo.toml` / `Cargo.lock` on its own.

this also means that existing PRs for new examples should not be merged
before this PR is merged so that they can be rebased and the config
updated in them (alternatively, if they're merged first then this PR
should be updated accordingly).

this solves rtic-rs/rtic-examples#44

[dependabot dependency version update]: https://docs.github.com/en/code-security/dependabot/dependabot-version-updates