automate `dependabot.yml` updating, add check to CI for it
to automate the process, a new script has been created to make life
easier in the future.
this script has also been included in the CI build and `git diff
--quiet` is used to check whether there are any modifications in the
local working directory (returns a failure status if any modifications
are present). this way it's not possible to forget adding a new example
to the dependabot config.
`git status` and `git diff` are also run in the workflow to make it more
obvious what the reason for the failure is in case there's a difference
(was very helpful for me now to debug the workflow because i learnt that
i had to sort the results of `find` as the order would otherwise be
arbitrary and on the CI it'd be different than locally, thus resulting
in a diff...).
the `LC_ALL=C` is required to ensure that the sorting is done the same
on all machines.